Project/Package_ISD/requirement.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
apipkg==1.5
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.6.1
colorama==0.4.0
comtypes==1.1.7
cycler==0.10.0
Cython==0.29.1
dask==0.20.0
decorator==4.3.0
defusedxml==0.5.0
et-xmlfile==1.0.1
execnet==1.5.0
idna==2.7
xlrd==1.2.0
imutils==0.5.2
jdcal==1.4
kiwisolver==1.0.1
lxml==4.2.5
more-itertools==4.3.0
multi-key-dict==2.0.3
natsort==5.4.1
networkx==2.2
nose==1.3.7
ntlm-auth==1.2.0
numpy==1.15.3
openpyxl==2.5.9
pbr==5.1.0
Pillow==5.3.0
py==1.7.0
pyasn1==0.4.4
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.3.0
pypng==0.0.18
pytesseract==0.2.5
pytest==5.3.2
pytest-forked==0.2
pytest-xdist==1.23.2
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.6
PyWavelets==1.0.1
requests==2.20.0
requests-ntlm==1.1.0
requests-oauthlib==1.0.0
responses==0.10.2
scikit-image==0.14.1
scipy==1.1.0
selenium==3.14.1
six==1.11.0
toolz==0.9.0
unicode==2.6
Unidecode==1.0.22
urllib3==1.24
Werkzeug==0.14.1
XlsxWriter==1.1.2
xmltodict==0.11.0